refactoring: rename jobs fields, fix services
